Test API endpoints.

+++ database/migrations/2020_03_11_235304_create_weathers_table.php
@@ -16,7 +16,7 @@ class CreateWeathersTable extends Migration
         Schema::create('weathers', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('flight_id');
-            $table->float('temperature');
+            $table->integer('temperature');
             $table->text('weather_blurb');
             $table->timestamps();
